Facilitate the use of GNU libquadmath if/when needed
Add commented out sections to quickly change to `__float128`.
Template and comment out `miegruneisen.cpp` as it interferes
and is not needed.
Add the ERK(4, 4) file to the project. It might also need
`LIBS += -lquadmath`.
Make `main.cpp` ready to work with 1D equations and
detonations, to be more specific.

+++ arithmeticwith.h
@@ -1,7 +1,10 @@
 #ifndef ARITHMETICWITH_H
 #define ARITHMETICWITH_H
 
+// #include <cmath>
 #include <concepts>
+
+// #include <quadmath.h>
 
 
 //template<class N>
@@ -41,7 +44,36 @@ concept ArithmeticWith = AddableWith<T, N>
 		&& DivisibleWith<T, N>;
 
 
-using numeric_val = double;
+using numeric_val = long double;
+// using numeric_val = double_t;
+// using numeric_val = __float128;
+
+
+//namespace std {
+////numeric_val cos(numeric_val qn) {
+////	return cosq(qn);
+////}
+
+
+////numeric_val sin(numeric_val qn) {
+////	return sinq(qn);
+////}
+
+
+//numeric_val sqrt(numeric_val qn) {
+//	return sqrtq(qn);
+//}
+
+
+//numeric_val pow(numeric_val qn1, numeric_val qn2) {
+//	return powq(qn1, qn2);
+//}
+
+
+////numeric_val exp(numeric_val qn) {
+////	return expq(qn);
+////}
+//}
 
 
 #endif // ARITHMETICWITH_H
